Fixed bug in OpenGL, GLviewport was getting set to neg numbers. Improved texture loader. clean up in tower and main. increased size of foundation.png. delete extra files in Data.
git-svn-id: https://svn.code.sf.net/p/highrisedev/code@106 f146f8d7-247d-460d-a1b3-d0a612191574

+++ Tower/floorbase.cpp
@@ -34,11 +34,11 @@ FloorBase::Draw ()
 }
 
 FloorBase::FloorBase (int x, int x2, int level, Tower * TowerParent)
-      :  mX (x)
-      ,  mX2 (x2)
+      :  mX ((float)x)
+      ,  mX2 ((float)x2)
       ,  mTowerParent (TowerParent)
 {
    mLevel = level;
-   mY = (mLevel * -36);
+   mY = (float)(mLevel * -36);
    mZ = -0.5f;
 }
